refactor: 노선 찾는 메소드 구현
